pyproject.toml
[tool.poetry]
name = "wss_multitools"
version = "0.1.0"
description = ""
authors = ["Ads <Adsis369@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
loguru = "^0.6.0"
zeep = "^4.1.0"
PyYAML = "^6.0"
xmltodict = "^0.12.0"
selenium = "^4.1.3"
SQLAlchemy = "^1.4.32"
httpx = "^0.22.0"
pyperclip = "^1.8.2"
pynput = "^1.7.6"
diagrams = "^0.21.1"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"
install = "^1.3.5"
tabulate = "^0.8.9"
pytest = "^7.1.2"
black = {extras = ["jupyter"], version = "^22.3.0"}
webdriver-manager = "^3.7.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pandas = "^1.4.1"
pylint = "^2.13.4"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
